requirements.txt
Django
psycopg2
djangorestframework-simplejwt
django-createsuperuser